SuperLU 4.3
SuperLU is a general purpose library for the direct solution of large,
sparse, nonsymmetric systems of linear equations. It is callable from
either C or Fortran.

Configuration is done in a separate customization file, which must be
modified by the brew script. It also originally bundles parts of COLAMD
and TMGLIB which should rather be installed by other Homebrew packages.

Tested on Mountain Lion 10.8.4 using Xcode 4.6.1.